Open for Discussion: Payouts in Claimable Balances vs. Direct Payments
So far, all payouts have been made using direct payment operations to
the addresses provided in `ADDRESSES.yml`. However, our `README` notes
that payments will be in the form of claimable balances.

Is there a "right" or "wrong" way to implement that? Which is the
preferred option for our use-case?

If direct payments are the way to go, here's a quick change that could
clarify it for hunters/reviewers.

Signed-off-by: Elliot J. Voris <elliot@voris.me>